move search engine types to its own file
https://coveord.atlassian.net/browse/KIT-2600
